Your program should process various user requests. In this stage, your program should:
Welcome users;
Display the instructions;
Ask for a request;
If a user enters zero, terminate the program;
If a user enters an empty request, print the instructions;
If numbers are not natural, print an error message;
If one number is entered, calculate and print the properties of this number;
For two numbers, print the list of numbers with properties;
Once the request is processed, continue execution from step 3.
In the current stage, the property names include even, odd, buzz , duck, palindromic and gapful. The test won't check the order of properties, their indentation, and spaces.